Move all Javascript in the head to the end of the body.
This prevents the browser from attempting to load Javascript before rendering the page, which should speed up the critical path and help the page start to render faster.  We have to leave Optimizely in place in order for it to be able to modify the page for A/B testing, but everything else can happen after the page is loaded.