Start of moving JavaScript out of the head.
We've moved all of the JavaScript, sans Optimizely, out of the head and into the bottom of the body.  Since some templates have arbitrary script tags they inject, which occur above the bottom of the body, we also introduced a new mechanism for Mako templates to declare JavaScript which can/should be loaded asynchronously, which will add it to the bottom of the body.