Include non-obsolete location info in student profile report
The student profile report that can be downloaded from the course staff
dashboard included the "location" field (which is obsolete) and the
mailing address, which is seldom completed. Here, we add the "country"
and "city" fields to the csv report. To do so, we need to be able to
dump the new fields to JSON so we convert the user fields to unicode
when needed.

Note that this breaks compatibility with earlier reports.