Add finished() metric for login and registration.
Call New Relic's finished() event marker to measure when the login and
registration pages actually become usable. This is important because the
pageLoad event can be distorted by analytics and marketing tools that
trigger after the page looks complete to the user.

Note: The "xfinished" PageAction is there to work around a New Relic bug
where "finished" events weren't making it into Insights. The reason we
have both is because finished() *does* still show up in session traces.

[PERF-298]