Remove unnecessary VideoModule save_states.
Before this commit, roughly half the save_state AJAX calls for the
VideoModule were done at load time, in order to send information about
YouTube availability (the value for which was almost always true).

This commit sends the value for what the LMS already thinks YouTube
availability is for this user, so that the AJAX callback is only
used in the case where the client side sees something different.

[PERF-262]