Add a link to the developer's guide under Dogwood
It's confusing that the developer's guide appears under Latest but not the Open edX release section of the docs landing page, particularly since Dogwood appears first. Let's add it above. It doesn't make sense for it to be versioned along with the Open edX release, so it just points to latest.