Allow to create users with a password hash instead of a clear text password.
The management command `create_user` currently only supports passing in a clear
text password.  This change introduces a flag to indicate that the password that
is passed in is a Django password hash instead. This allows to import users from
different Django apps, and to avoid having clear text passwords in the Ansible
variables.