Proof of concept
I was able to reproduce the hotfix #2 error locally by changing the
task to use update_collected() once more, then publishing a change in a course
containing a video module. The cms celery worker crashed on this task.

The naive change made to dump the task onto an lms worker succeeded.