Clarify account lockout messaging
Edit the account lockout message so users know the lockout time.
And overestimate the lockout time in the message by 5 minutes becasue
if a user trys to login at 14 minutes 59 seconds then the lockout
timer is reset for 15 more minutes.  Also make the user understand
that reseting their password will not reset the lockout.