Serve localfs reports over HTTP
When using localfs to serve instructor reports, do it over HTTP, as
opposed to via "file://".  This allows for running devstack on
containers, and also for constrained production use (where the user is
guaranteed to hit the same web server on multiple requests).

NOTE: this has been superseded by the new Django storage backend
introduced in 2faab97e.