Retry-while-retrying is not an error in bulk email (#13068)
This code will be hit if, during a retry of the send_email task, an error
is encountered that means we need to retry the task again. Since the task will
be retried with timed backoff up to predefinied limits, this is an expected
situation and does not need to be logged as an exception.