Fix grade push back for non-components through LTI
There is a bug introduced in this commit:
https://github.com/edx/edx-platform/commit/22046d4067595f785365554b430fad1fc9d85000#diff-32687833c38e231b51a8f27a76c72a56R119
The return value of the locations_to_weighted_scores[location] is
changed from `Score` object to a tuple (Score, weight). However, the
left side of assignment didn't get updated. So an error is raised and
the Celery task failed to continue. Thus, no grade is being passed back.