Disambiguate queryset extensions for ConfigurationModels
This change disambiguates the use of `id` in the
ConfigurationModelManager queryset extensions. This is required
when specifying KEY_FIELDS where there is more than one key and
at least one of them is a ForeignKey. The table must be specified
in this case because the query extension may be appended to the
WHERE clause of queries selecting `id` fields from multiple tables
via an INNER JOIN clause.