Modify csv downloads in instructor dashboard to extract student answer from JSON in some problem types
The 'Download a CSV of problem responses' button in the Data Download tab of the instructor dashboard
used to produce 2 columns: 'username' and 'state', where the latter is a hard-to-read json object. This
commit extracts the student's answer in submit-and-compare and free-text-response problems and replaces
the usual 'state' with this answer.