Disable migrations by default in unittests
Rendering model states when running migrate in Django 1.8 consumes
on the order of 4x-6x more memory than <1.8.

References:
* https://code.djangoproject.com/ticket/24745
* https://code.djangoproject.com/ticket/24743

This is causing failures in Jenkins as the workers run out of memory
and either processes cannot fork or the Out Of Memory Killer
kills a process that is running tests.

Disabling migrations until we do either or both of these:
* Upgrade to Django 1.9 and turn them back on. See EV-92 "Discovery: Upgrade edx-platform to Django 1.9"
* Address TE-1310 "Unit tests should be able to run without migrations"