Fix for spurious 'cb() never called' errors.
Some Jenkins builds are erroring when attempting to set up the
npm environment with the error:

`npm ERR! cb() never called!`

Although it could point to a couple of different things, it is more classically
a sign of a dirty npm environment cache. Cleaning the cache at the start
of any given Jenkins job (which are sourced through this file) should
alleviate the problem.