Removes an redundant call to loadDiscussionApp()
This function is already called from discussion_forum.js on page load, and so
calling it again was causing the click handlers to be doubled up.

E.g., if you click on the "All Discussions" button in DiscussionCourseXBlock,
it would toggle the button twice, thus never displaying the "All Discussions"
list.