Allow Stanford configuration to list the problem types Instructors can query
In the Queries tab of the Instructor dashboard, instructors can query problems and sections
of their courses by the type of problem and the different sections. This commit, coupled
with a change in the configuration repo, allows Stanford to specify which problem types
instructors can query by. With the corresponding configuration commit, we would now include
peer assessment, submit and compare, free-text response, MUFI, survey, inline-dropdown, poll
and image-coding.