improved business logic around bulk enrollment
For courses with a professional course mode, remove the entire bulk
enrollment section from the "Membership" tab of the Instructor
Dashboard (see screenshot attached) for non-staff users.

For courses without a professional course mode, remove the "Select a
course mode" button from the bulk enrollment section for non-staff
users and use "audit" as the course mode when bulk enrolling users.