Spike: Linear SQL query count by subsections.
Run the following:

$ SEQ=60 paver test_system -t lms/djangoapps/grades/tests/test_tasks.py

for any number of SEQ between 1 and 76.

When updating scores, the number of queries increases linearly.

Tests will pass, as per this commit, because there are 2 extra queries
per sequence.  For scalability, this should be a constant number of
queries.  Above 76, there is 1 extra query beyond what is expected.  At
SEQ=0 (which still has one sequence), there are 3 fewer queries than
expected.

Logged as TNL-6225