Remove Datadog instrumentation around model saves
We added this a long time ago as part of the shift to opaque-keys, but
left it on because it's occasionally useful for spotting unusual levels
of activity for certain models. That being said, because we have so
many models now, we're generating too many differently named metrics,
which might get expensive.

This is doing the simplest thing and just removing datadog reporting
of model updates altogether.