Disable changes to course code in re-run form
ECOM-6898

Disable (make read-only) the field for course code in the re-run form.  This will prevent creation of course runs with a different course code in the course key that are intended to stand for the same course in a program, which presents problems for the course catalog mapping process.