Removes an obsolete Django setting, ALLOW_ALL_ADVANCED_COMPONENTS.
We moved from configuring installed components in settings.py files to using just Python entry points.
This used to be a setting which enabled all XBlocks with appropriate entry points.
It was removed from the rest of the codebase in a previous PR, and this usage was
inadvertently left behind.