Remove dealer.git use for LMS.
We use the dealer.git library to determine the current revision of
edx-platform at startup time and store that as EDX_PLATFORM_REVISION in
our settings. This can take a surprisingly long time and involve a lot
of disk IO to initialize.

Studio still needs this functionality because it determines its static
asset folder based partly on the hash. This is horrible, because it
means all static assets are invalidated on every release instead of
just the ones that change. But that will be the subject of a different
PR.