LMS startup optimization: Load pynliner locally.
pynliner imports cssutils, which has an expensive initialization. All
told, it can account for 15-20% of "fast" LMS startup (without asset
compilation). So we're going to load it locally here so that we delay
that one-time hit until we actually do the (rare) operation that is
sending a credit notification email.