Fixes custom theme permissions
Gunicorn runs as the www-data and when it creates the file it is owned
by www-data:www-data.... If however we do any work from the management
console we are doing so as the edxapp user and won't be able to change
that file (when calling .save() on the SiteConfiguration model).
Hence the 777 permission. It's just a CSS file so it doesn't
matter, and not any user can get into that nested directory at all, so
we should be fine.