Auto-advance video toggle
This adds an auto-advance toggle to the video player.  If turned on,
as soon as the video ends the next unit or subsection will be loaded.

Since this is a global preference, when it is set loading a section with
a video will automatically start it.  This has the (desired) effect that
consecutive video sections will play automatically from start to finish.