Don't allow any user to create courses by default
In f095c5fec6507ba38552b0e02530e8ce981ffcbe it was decided to enable
anyone to make courses in Studio by default in order to make things
easier in dev environments. This was before the code was open source.
This default is likely in effect for many Open edX instances, but almost
no one would want this setting outside of development.