Allowing writing of reviews on reviews page.
LEARNER-1628

This allows enrolled users to write reviews on the reviews page that
can be accessed through the course tools sidebar. A button in the top
corner of the page toggles the widget that is configured with a
setting in the common.py file. The reviews tool is Course Talk.