UX improvements on learner profile.
LEARNER-1857

This story constitutes a large change to the UX of the learner profile as well as the deprecation of the Backbone.js implementation in favor of a Django Fragment implementation.

Some specific changes include adding a flexible learner profile card, limiting the bio entry to only 300 characters, adding social links (not currently functional for this story), adding a joined date for the user (not currently functional for this story) and assuring that the page is responsive.