Re-add to BlockStructure cache on miss.
BlockStructures are stored in S3, but they're cached in memcached. Prior
to this commit, they were only written to memcached when first created.
This commit adds a write to repopulate the cache entry every time there
is a cache miss. Hopefully, this will help reduce the miss rate we're
seeing.