Create blank footer style for responsive theming
In the push to make pages more responsive, we need to replace the
header and footer in our WL sites to be used with both bootstrap
and non bootstrap pages. Currently, the footer in LMS already
responsive, so this change is only to allow for comprehensive
themed sites to override them. Eventually, if the LMS footer needs
to be rewritten, it may use these files.

Part of [WL-1286]

Based off of Harry's work for the header in PR #16210