Avoid a django breaking change: set_language response code
In Django 1.10+, the set_language view (/i18n/setlang) will respond to
Ajax requests which do not contain a "next" parameter with status 204 No
Content instead of 304.  This commit adds the "next" parameter to the
request in order to ensure that upgrading to Django 1.10+ will not cause
the set_language view to change behavior.

PLAT-1353