Show current tab even in high contrast mode
In the dashboard header and the course navigation bar, the current
tab is supposed to be the only one with a bottom border.  But in
high contrast mode in Windows, all of them were assigned a bottom
border color.

So instead of toggling the color when active, toggle the border
style, since Windows overrides our color choice anyway.

LEARNER-3219