Remove dealer.git (startup optimization).
In the absence of an EDX_PLATFORM_VERSION OS env variable, we used to
calculate the current version by looking at the current git revision of
the edx-platform repo. This was an expensive calculation (~7% of startup
costs). It also turns out that at some point we stopped setting that env
variable anywhere, so it always fell back to recalculating it.

With this commit, we'll start reading the code version from the env JSON
files, and fall back to "unknown" rather than doing the dealer.git call.