Ignore certain patterns during collectstatic.
There are symlinks and file types that are not necessary to copy over
to the actual static files output, either because they are source files
that get compiled out, or because they exist only for tests. FEDX-448.

With themes enabled, this reduces the unoptimized static asset build
size from 343M to 212M. The optimized version goes from 838M to 500M.