Improve search error messages
When a more specific error message is available from the search backend,
show the user that instead of a blanket "There was an error, try
searching again."

In github.com/edx/edx-search, we're proposing removing quotes from the
list of reserved characters in Elastic search terms. This enables
searching for exact phrases, but also introduces the possibility of
malformed queries. In those cases, it would be nice to give the user a
hint that they need to change their query, instead of simply trying the
same one again.