Maintenance banner functionality, including a new waffle switch and setting
This adds the following new waffle switch (and namespace):

open_edx_util.display_maintenance_warning

and a new setting:

settings.MAINTENANCE_BANNER_TEXT

The provided view decorator allows developers to select which views
receive the static banner, and the banner can be globally toggled via
the waffle switch.  The text of the banner is determined by the setting,
which is currently not translated.